Создает контроллер комментариев WIP
